requirements.txt
crewai>=0.95.0
fastapi>=0.115.0
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.58.0
tavily-python>=0.5.0
uvicorn>=0.30.0
weasyprint>=63.0
python-dotenv>=1.0.1
pydantic>=2.8.0
beautifulsoup4>=4.12.0
requests>=2.32.0
pytest>=8.0.0
